# docs/swagger.yaml
basePath: /api/v0
definitions:
...
  core.EnvMap:
    additionalProperties:
      type: string
    type: object
...
  core.Input:
    additionalProperties: {}
    type: object
...
  router.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
...
  router.PageInfoDTO:
    properties:
      limit:
        example: 50
        type: integer
      next_cursor:
        example: v2:after:tool-001
        type: string
      prev_cursor:
        example: v2:before:tool-000
        type: string
      total:
        example: 2
        type: integer
    type: object
  router.ProblemDocument:
    properties:
      code:
        example: invalid_cursor
        type: string
      detail:
        example: Invalid cursor parameter
        type: string
      instance:
        example: /api/v0/workflows
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
      type:
        example: about:blank
        type: string
    type: object
  router.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/router.ErrorInfo'
      message:
        type: string
      status:
        type: integer
    type: object
...
  task.Type:
    enum:
      - basic
      - router
      - parallel
      - collection
      - aggregate
      - composite
      - signal
      - wait
      - memory
    type: string
    x-enum-varnames:
      - TaskTypeBasic
      - TaskTypeRouter
      - TaskTypeParallel
      - TaskTypeCollection
      - TaskTypeAggregate
      - TaskTypeComposite
      - TaskTypeSignal
      - TaskTypeWait
      - TaskTypeMemory
...
  tkrouter.TaskDTO:
    properties:
      action:
        type: string
      condition:
        type: string
      env:
        $ref: '#/definitions/core.EnvMap'
      has_subtasks:
        type: boolean
      id:
        type: string
      items: {}
      mode:
        type: string
      outputs:
        $ref: '#/definitions/core.Input'
      routes:
        additionalProperties: {}
        type: object
      signal_name:
        type: string
      strategy:
        type: string
      subtask_ids:
        items:
          type: string
        type: array
      timeout:
        type: string
      type:
        $ref: '#/definitions/task.Type'
      with:
        $ref: '#/definitions/core.Input'
    type: object
  tkrouter.TaskListItem:
    properties:
      action:
        type: string
      condition:
        type: string
      env:
        $ref: '#/definitions/core.EnvMap'
      etag:
        example: abc123
        type: string
      has_subtasks:
        type: boolean
      id:
        type: string
      items: {}
      mode:
        type: string
      outputs:
        $ref: '#/definitions/core.Input'
      routes:
        additionalProperties: {}
        type: object
      signal_name:
        type: string
      strategy:
        type: string
      subtask_ids:
        items:
          type: string
        type: array
      timeout:
        type: string
      type:
        $ref: '#/definitions/task.Type'
      with:
        $ref: '#/definitions/core.Input'
    type: object
  tkrouter.TaskResponse:
    properties:
      action:
        type: string
      condition:
        type: string
      env:
        $ref: '#/definitions/core.EnvMap'
      has_subtasks:
        type: boolean
      id:
        type: string
      items: {}
      mode:
        type: string
      outputs:
        $ref: '#/definitions/core.Input'
      routes:
        additionalProperties: {}
        type: object
      signal_name:
        type: string
      strategy:
        type: string
      subtask_ids:
        items:
          type: string
        type: array
      timeout:
        type: string
      type:
        $ref: '#/definitions/task.Type'
      with:
        $ref: '#/definitions/core.Input'
    type: object
  tkrouter.TasksListResponse:
    properties:
      page:
        $ref: '#/definitions/router.PageInfoDTO'
      tasks:
        items:
          $ref: '#/definitions/tkrouter.TaskListItem'
        type: array
    type: object
...
info:
  contact:
    email: support@compozy.com
    name: Compozy Support
    url: https://github.com/compozy/compozy
  description: Compozy is a Next-level Agentic Orchestration Platform, tasks, and tools. Authentication
    is configurable via the SERVER_AUTH_ENABLED environment variable. When enabled, all endpoints require
    Bearer token authentication. When disabled (typical for development), endpoints are accessible without
    authentication.
  license:
    name: BSL-1.1
    url: https://github.com/compozy/compozy/blob/main/LICENSE
  termsOfService: https://github.com/compozy/compozy
  title: Compozy API
  version: "1.0"
paths:
...
  /tasks:
    get:
      consumes:
        - application/json
      description: List tasks with cursor pagination. Optionally filter by workflow usage.
      parameters:
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
        - description: Return only tasks referenced by the given workflow
          example: '"wf1"'
          in: query
          name: workflow_id
          type: string
        - description: Page size (max 500)
          example: 50
          in: query
          name: limit
          type: integer
        - description: Opaque pagination cursor
          in: query
          name: cursor
          type: string
        - description: Filter by task ID prefix
          in: query
          name: q
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Tasks retrieved
          headers:
            Link:
              description: RFC 8288 pagination links for next/prev
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/tkrouter.TasksListResponse'
                type: object
        "400":
          description: Invalid cursor
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "404":
          description: Workflow not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: List tasks
      tags:
        - tasks
  /tasks/{task_id}:
    delete:
      description: Delete a task configuration. Returns conflict when referenced.
      parameters:
        - description: Task ID
          example: '"approve-request"'
          in: path
          name: task_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
      produces:
        - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "409":
          description: Task referenced
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Delete task
      tags:
        - tasks
    get:
      consumes:
        - application/json
      description: Retrieve a task configuration by ID.
      parameters:
        - description: Task ID
          example: '"approve-request"'
          in: path
          name: task_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Task retrieved
          headers:
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/tkrouter.TaskDTO'
                type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Get task
      tags:
        - tasks
    put:
      consumes:
        - application/json
      description: Create a task configuration when absent or update an existing one using strong ETag
        concurrency.
      parameters:
        - description: Task ID
          example: '"approve-request"'
          in: path
          name: task_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
        - description: Strong ETag for optimistic concurrency
          example: '"\"abc123\""'
          in: header
          name: If-Match
          type: string
        - description: Task configuration payload
          in: body
          name: payload
          required: true
          schema:
            additionalProperties: true
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: Task updated
          headers:
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/tkrouter.TaskDTO'
                type: object
        "201":
          description: Task created
          headers:
            Location:
              description: Absolute URL for the task
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/tkrouter.TaskDTO'
                type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "409":
          description: Task referenced
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "412":
          description: ETag mismatch
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Create or update task
      tags:
        - tasks
...
  /workflows/{workflow_id}/tasks:
    get:
      consumes:
        - application/json
      description: Retrieve a list of all tasks within a specific workflow
      parameters:
        - description: Workflow ID
          example: '"data-processing"'
          in: path
          name: workflow_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Tasks retrieved successfully
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    properties:
                      tasks:
                        items:
                          $ref: '#/definitions/tkrouter.TaskResponse'
                        type: array
                    type: object
                type: object
        "400":
          description: Invalid workflow ID
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
        "404":
          description: Workflow not found
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
      summary: List tasks for a workflow
      tags:
        - tasks
  /workflows/{workflow_id}/tasks/{task_id}:
    get:
      consumes:
        - application/json
      description: Retrieve a specific task configuration by its ID within a workflow
      parameters:
        - description: Workflow ID
          example: '"data-processing"'
          in: path
          name: workflow_id
          required: true
          type: string
        - description: Task ID
          example: '"validate-input"'
          in: path
          name: task_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Task retrieved successfully
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/tkrouter.TaskResponse'
                type: object
        "400":
          description: Invalid workflow or task ID
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
        "404":
          description: Task not found
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
      summary: Get task by ID
      tags:
        - tasks
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication (Optional - controlled by SERVER_AUTH_ENABLED environment
      variable. When enabled, all API endpoints require valid authentication. In development/testing,
      authentication may be disabled.)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...